Add inline background color to HTML (#a79e9c)
I added a background color to the HTML element using the inline style attribute. The color #a79e9c was chosen because it’s a soft, neutral tone that blends well with the rest of the page design. By applying this directly in the HTML, I made sure the color shows up consistently across different browsers and devices.